requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.23,<1.0
python-multipart>=0.0.9,<0.1.0
pillow>=10.0,<11.0
tensorflow-cpu>=2.15,<2.17
numpy>=1.26,<3.0
streamlit>=1.36,<2.0

